ScanBfs, QueueBfs, added

+++ src/graph.cpp
@@ -59,8 +59,8 @@ void Graph::normGraph() {
 
 void Graph::convGraph() {
     for (int i = 0; i <= n; i++) {
-        posV.emplace_back(packE.size());
-        packE.emplace_back(E[i].size());
+        posV.emplace_back((int)packE.size());
+        packE.emplace_back((int)E[i].size());
         packW.emplace_back(0);
         for (auto it : E[i]) {
             packE.emplace_back(it.first);
